Add files via upload
Initial Upload of raw files for printing/editing. No internals or non-printed parts included in this commit. More to come depending on desire - I'm not really in the mood to add more than this right now and it should be enough to start people off. If you want the Wildcard STEP files download them from Slug's repository.